eth, les: move consensus engine creation to common package
